Puting all structs in new package httpcheck
