+++ beeping.go
@@ -5,17 +5,16 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"flag"
-	"fmt"
 	"io/ioutil"
 	"log"
 	"net"
 	"net/http"
 	"net/http/httptrace"
-	"net/url"
 	"os"
 	"strings"
 	"time"
 
+	"github.com/olpia/beeping/httpcheck"
 	"github.com/oschwald/geoip2-golang"
 	"github.com/tcnksm/go-httpstat"
 	"github.com/yanc0/beeping/sslcheck"
@@ -41,50 +40,6 @@ type Beeping struct {
 	Message string `json:"message"`
 }
 
-// Check defines the check to do
-type Check struct {
-	URL      string        `json:"url" binding:"required"`
-	Pattern  string        `json:"pattern"`
-	Header   string        `json:"header"`
-	Insecure bool          `json:"insecure"`
-	Timeout  time.Duration `json:"timeout"`
-	Auth     string        `json:"auth"`
-}
-
-type Timeline struct {
-	NameLookup    int64 `json:"name_lookup"`
-	Connect       int64 `json:"connect"`
-	Pretransfer   int64 `json:"pretransfer"`
-	StartTransfer int64 `json:"starttransfer"`
-}
-
-type Geo struct {
-	Country string `json:"country"`
-	City    string `json:"city,omitempty"`
-	IP      string `json:"ip"`
-}
-
-// Response defines the response to bring back
-type Response struct {
-	HTTPStatus      string `json:"http_status"`
-	HTTPStatusCode  int    `json:"http_status_code"`
-	HTTPBodyPattern bool   `json:"http_body_pattern"`
-	HTTPHeader      bool   `json:"http_header"`
-	HTTPRequestTime int64  `json:"http_request_time"`
-
-	InstanceName string `json:"instance_name"`
-
-	DNSLookup        int64 `json:"dns_lookup"`
-	TCPConnection    int64 `json:"tcp_connection"`
-	TLSHandshake     int64 `json:"tls_handshake,omitempty"`
-	ServerProcessing int64 `json:"server_processing"`
-	ContentTransfer  int64 `json:"content_transfer"`
-
-	Timeline *Timeline          `json:"timeline"`
-	Geo      *Geo               `json:"geo,omitempty"`
-	SSL      *sslcheck.CheckSSL `json:"ssl,omitempty"`
-}
-
 func NewErrorMessage(message string) *ErrorMessage {
 	var response = ErrorMessage{}
 	response.Message = message
@@ -95,59 +50,14 @@ func InvalidJSONResponse() *ErrorMessage {
 	return NewErrorMessage("Invalid JSON sent")
 }
 
-func NewResponse() *Response {
-	var response = Response{}
-	response.Timeline = &Timeline{}
+func NewResponse() *httpcheck.Response {
+	var response = httpcheck.Response{}
+	response.Timeline = &httpcheck.Timeline{}
 	return &response
 }
 
-func NewCheck() *Check {
-	return &Check{Timeout: 10}
-}
-
-// Performs some validation checks on the target.
-// Returns nil if valid, returns an error otherwise.
-func (check *Check) validateTarget() error {
-	targetURL, err := url.Parse(check.URL)
-	if err != nil {
-		return err
-	}
-	ip := net.ParseIP(targetURL.Hostname())
-	if ip == nil {
-		// Hostname provided is not an IP. Without whitelisting, it is not possible to tell
-		// whether it is an internal hostname.
-		return nil // For now, hostnames are not needed for this check.
-	}
-
-	// Check for local network IPs
-	switch {
-	// Loopback address
-	case ip.IsLoopback():
-		return fmt.Errorf("Disallowed target")
-	// Link-local unicast
-	case ip.IsLinkLocalUnicast():
-		return fmt.Errorf("Disallowed target")
-	// Link-local multicast
-	case ip.IsLinkLocalMulticast():
-		return fmt.Errorf("Disallowed target")
-	// Private network (10.0.0.0/8)
-	case len(ip) == 4 && ip[0] == 10:
-		return fmt.Errorf("Disallowed target")
-	// Private network (Carrier-grade NAT; 100.64.0.0/10)
-	case len(ip) == 4 && ip[0] == 100 && ip[1] >= 64 && ip[1] <= 127:
-		return fmt.Errorf("Disallowed target")
-	// Private network (172.16.0.0/12)
-	case len(ip) == 4 && ip[0] == 172 && ip[1] >= 16 && ip[1] <= 31:
-		return fmt.Errorf("Disallowed target")
-	// Private network (192.168.0.0/16)
-	case len(ip) == 4 && ip[0] == 192 && ip[1] == 16:
-		return fmt.Errorf("Disallowed target")
-	// Private network (fc00::/7)
-	case len(ip) == 16 && (ip[0] == 0xfc || ip[0] == 0xfd):
-		return fmt.Errorf("Disallowed target")
-	}
-
-	return nil
+func NewCheck() *httpcheck.Check {
+	return &httpcheck.Check{Timeout: 10}
 }
 
 func main() {
@@ -206,7 +116,7 @@ func handlerCheck(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if err := check.validateTarget(); err != nil {
+	if err := check.ValidateTarget(); err != nil {
 		log.Println("[WARN] Invalid target:", err.Error())
 		jsonRes, _ := json.Marshal(NewErrorMessage(err.Error()))
 		w.Write(jsonRes)
@@ -227,7 +137,7 @@ func handlerCheck(w http.ResponseWriter, r *http.Request) {
 }
 
 // CheckHTTP do HTTP check and return a beeping response
-func CheckHTTP(check *Check) (*Response, error) {
+func CheckHTTP(check *httpcheck.Check) (*httpcheck.Response, error) {
 	var response = NewResponse()
 	var conn net.Conn
 
@@ -347,7 +257,7 @@ func milliseconds(d time.Duration) int64 {
 	return d.Nanoseconds() / 1000 / 1000
 }
 
-func geoIPCountry(geodatabase string, ip string, response *Response) error {
+func geoIPCountry(geodatabase string, ip string, response *httpcheck.Response) error {
 	db, err := geoip2.Open(geodatabase)
 	if err != nil {
 		return err
@@ -359,7 +269,7 @@ func geoIPCountry(geodatabase string, ip string, response *Response) error {
 	if err != nil {
 		return err
 	}
-	response.Geo = &Geo{}
+	response.Geo = &httpcheck.Geo{}
 	response.Geo.Country = record.Country.IsoCode
 	response.Geo.IP = ip
 	if record.Country.Names != nil {
@@ -368,7 +278,7 @@ func geoIPCountry(geodatabase string, ip string, response *Response) error {
 	return nil
 }
 
-func instanceName(name string, response *Response) error {
+func instanceName(name string, response *httpcheck.Response) error {
 	var err error
 	response.InstanceName = name
 	if name == "" {
